completed till calling authorization api

+++ src/app/user/shared/security/interceptors/auth-config.interceptor.ts
@@ -5,26 +5,29 @@ import { Router } from '@angular/router';
 import { inject } from '@angular/core';
 
 export const authConfigInterceptor: HttpInterceptorFn = (req, next) => {
-
-  const authService = inject(AuthService);
-  const router = inject(Router);
-
-  // Check if user is logged in but not authorized and redirect if needed
-  if (authService.isLoggedIn() && !authService.isUserAuthorized()) {
-    authService.removeToken();
-    router.navigate(['/']);
-    return throwError(() => new Error('Unauthorized access')); // Throw error to stop request chain
+  
+  const url: string = req.url;
+  if (!(url.includes('accounts.spotify.com') || url.includes('api.spotify.com'))) {
+    const authService = inject(AuthService);
+    const router = inject(Router);
+  
+    // Check if user is logged in but not authorized and redirect if needed
+    if (authService.isLoggedIn() && !authService.isUserAuthorized()) {
+      authService.removeToken();
+      router.navigate(['/']);
+      return throwError(() => new Error('Unauthorized access')); // Throw error to stop request chain
+    }
+  
+    // Add authorization header if token exists
+    const authToken = authService.getAccessToken();
+    if (authToken) {
+      req = req.clone({
+        setHeaders: {
+          Authorization: `Bearer ${authToken}`,
+        },
+      });
+    } 
   }
-
-  // Add authorization header if token exists
-  const authToken = authService.getUserToken();
-  if (authToken) {
-    req = req.clone({
-      setHeaders: {
-        Authorization: `Bearer ${authToken}`,
-      },
-    });
-  }
-
+  
   return next(req);
 };
